feat: Added a lot of comments to RBTree fix: some changes in packages
