+++ BSTrees/src/main/kotlin/dataBases/SerializableType.kt
@@ -1,4 +1,4 @@
-package serialization
+package dataBases
 
 import kotlinx.serialization.Serializable
