+++ BSTrees/src/main/kotlin/dataBases/reps/JsonTreeRepo.kt
@@ -1,8 +1,8 @@
-package serialization.reps
+package dataBases.reps
 
 import kotlinx.serialization.decodeFromString
 import kotlinx.serialization.json.Json
-import serialization.SerializableTree
+import dataBases.SerializableTree
 import kotlinx.serialization.encodeToString
 import java.io.File
 import java.io.FileNotFoundException
