+++ BSTrees/src/main/kotlin/trees/binarySearch/BSTree.kt
@@ -1,25 +1,54 @@
-package binarySearchTree
+package trees.binarySearch
 
-import BTree
+import trees.BTree
 import kotlin.random.Random
 
+/**
+ * A class representing a randomized binary search tree.
+ *
+ *
+ * @generic <K> the type of key stored in the tree. It must be comparable
+ * @generic <V> the type of value stored in the tree
+ */
 class BSTree<K : Comparable<K>, V> : BTree<K, V, BSNode<K, V>>() {
 
+    /**
+     * A balancer class providing rotations of the tree with updating the size
+     */
     private val balancer = BSBalancer<K, V>()
 
+    /**
+     * Insert a node to the tree.
+     * It is actually a wrapper for the add function. Necessary to implement recursion
+     *
+     * @param value the value to add
+     * @param key the key under which the value is stored
+     */
     override fun insert(key: K, value: V) {
         add(BSNode(key, value))
     }
 
+    /**
+     * Add a node to the root of the tree.
+     * It's necessary to implement random adding.
+     *
+     * @param node the node to add
+     */
     private fun addRoot(node: BSNode<K, V>): BSNode<K, V> {
-
+        //temp is needed to avoid possibility of changing the root
         val temp = this.root
         return if (temp == null) node
         else if (temp.key == node.key) {
             temp.value = node.value
             temp
         } else {
+            /**
+             * Some kind of recursive implementation
+             * Creating subtree to execute add function again
+             * Then we return the node to the desired son
+             */
             val subTree = BSTree<K, V>()
+            //We make the usual adding, then we put the node in the root by rotation it tom the top
             if (node.key < temp.key) {
                 subTree.root = temp.leftNode
                 this.root?.leftNode = subTree.addRoot(node)
@@ -33,8 +62,13 @@ class BSTree<K : Comparable<K>, V> : BTree<K, V, BSNode<K, V>>() {
 
     }
 
+    /**
+     * Add a node to the tree
+     *
+     * @param node the node to add
+     */
     private fun add(node: BSNode<K, V>) {
-
+        //temp is needed to avoid possibility of changing the root
         val temp = this.root
         if (temp == null) this.root = node
         else if (temp.key == node.key) this.root?.value = node.value
@@ -44,9 +78,18 @@ class BSTree<K : Comparable<K>, V> : BTree<K, V, BSNode<K, V>>() {
             /*
             Randomized insertion into the root of the tree allows you
             to artificially balance it with a fairly small height.
+            The idea is that inserting into the root shuffles the tree well, which provides a relatively small height.
+            The tree is not perfectly balanced, but differs from the logarithm on average by no more than twice,
+            which is a constant
+
             Otherwise, we perform the usual insertion
             */
             else {
+                /**
+                 * Some kind of recursive implementation
+                 * Creating subtree to execute add function again
+                 * Then we return the node to the desired son
+                 */
                 val subTree = BSTree<K, V>()
                 if (temp.key > node.key) {
                     subTree.root = temp.leftNode
@@ -61,16 +104,27 @@ class BSTree<K : Comparable<K>, V> : BTree<K, V, BSNode<K, V>>() {
                 }
             }
 
+            //Updating this node before leaving recursion
             this.root?.updateSize()
 
         }
     }
 
+    /**
+     * Merge two nodes into one.
+     * One of the ways to implement removal from the tree
+     *
+     * @param left
+     * @param right
+     * The nodes to merge
+     */
     private fun join(left: BSNode<K, V>?, right: BSNode<K, V>?): BSNode<K, V>? {
 
+        //If one of the nodes is null then we can just return the second one
         if (left == null) return right
         if (right == null) return left
 
+        //Choosing the node that would be a new root
         return if (Random.nextInt() % (left.size + right.size) < left.size) {
             left.rightNode = join(left.rightNode, right)
             left.rightNode?.parent = left
@@ -85,6 +139,11 @@ class BSTree<K : Comparable<K>, V> : BTree<K, V, BSNode<K, V>>() {
 
     }
 
+    /**
+     * Delete a value from the tree.
+     *
+     * @param key the key under which the value is stored
+     */
     override fun delete(key: K) {
 
         val temp = this.root
@@ -106,6 +165,7 @@ class BSTree<K : Comparable<K>, V> : BTree<K, V, BSNode<K, V>>() {
                 this.root?.rightNode = subTree.root
             }
 
+            //Updating this node before leaving recursion
             this.root?.updateSize()
 
         }
