+++ BSTrees/src/test/kotlin/avlTree/AVLTreeTest.kt
@@ -4,6 +4,8 @@ import org.junit.jupiter.api.*
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
 import treeInvariants.TreesInvariants
+import trees.avl.AvlNode
+import trees.avl.AvlTree
 import kotlin.random.Random
 
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
