+++ BSTrees/src/test/kotlin/binarySearchTree/BSTreeTest.kt
@@ -4,6 +4,8 @@ import org.junit.jupiter.api.*
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
 import treeInvariants.TreesInvariants
+import trees.binarySearch.BSNode
+import trees.binarySearch.BSTree
 import kotlin.random.Random
 
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
